Second stage completed and running successfully.
